requirements.txt
allure-pytest
pytest
selene==2.0.0rc2
python-dotenv==0.20.0